test (test responder manager): pass mock responder as an argument
