+++ webapi/src/test/scala/org/knora/webapi/e2e/v1/SipiV1E2ESpec.scala
@@ -23,22 +23,26 @@ package org.knora.webapi.e2e.v1
 import java.io.File
 import java.nio.file.{Files, Paths}
 
+import akka.actor.ActorDSL._
 import akka.actor._
 import akka.pattern._
 import akka.util.Timeout
 import org.knora.webapi.e2e.E2ESpec
 import org.knora.webapi.messages.v1respondermessages.resourcemessages.{CreateResourceApiRequestV1, CreateResourceValueV1}
+import org.knora.webapi.messages.v1respondermessages.sipimessages._
 import org.knora.webapi.messages.v1respondermessages.triplestoremessages.{RdfDataObject, ResetTriplestoreContent}
-import org.knora.webapi.messages.v1respondermessages.valuemessages.{CreateFileV1, CreateRichtextV1}
+import org.knora.webapi.messages.v1respondermessages.valuemessages.{CreateFileV1, CreateRichtextV1, StillImageFileValueV1}
 import org.knora.webapi.responders._
 import org.knora.webapi.responders.v1._
 import org.knora.webapi.routing.v1.ResourcesRouteV1
 import org.knora.webapi.store._
+import org.knora.webapi.util.ActorUtil._
+import org.knora.webapi.util.InputValidation
 import org.knora.webapi.{FileWriteException, LiveActorMaker}
 import spray.http._
 
-import scala.concurrent.Await
 import scala.concurrent.duration._
+import scala.concurrent.{Await, Future}
 
 /**
   * End-to-end test specification for the resources endpoint. This specification uses the Spray Testkit as documented
@@ -52,7 +56,62 @@ class SipiV1E2ESpec extends E2ESpec {
          # akka.stdout-loglevel = "DEBUG"
         """.stripMargin
 
-    val responderManager = system.actorOf(Props(new TestResponderManagerV1), name = RESPONDER_MANAGER_ACTOR_NAME)
+    /**
+      * Imitates the Sipi server by returning a [[SipiResponderConversionResponseV1]] representing an image conversion request.
+      *
+      * @param conversionRequest the conversion request to be handled.
+      * @return a [[SipiResponderConversionResponseV1]] imitating the answer from Sipi.
+      */
+    private def imageConversionResponse(conversionRequest: SipiResponderConversionRequestV1): Future[SipiResponderConversionResponseV1] = {
+
+        // delete tmp file (depending on the kind of request given: only necessary if Knora stored the file - non GUI-case)
+        def deleteTmpFile(conversionRequest: SipiResponderConversionRequestV1): Unit = {
+            conversionRequest match {
+                case (conversionPathRequest: SipiResponderConversionPathRequestV1) =>
+                    // a tmp file has been created by the resources route (non GUI-case), delete it
+                    InputValidation.deleteFileFromTmpLocation(conversionPathRequest.source)
+                case _ => ()
+            }
+        }
+
+        val originalFilename = conversionRequest.originalFilename
+        val originalMimeType: String = conversionRequest.originalMimeType
+
+        val fileValuesV1 = Vector(StillImageFileValueV1(// full representation
+            internalMimeType = "image/jp2",
+            originalFilename = originalFilename,
+            originalMimeType = Some(originalMimeType),
+            dimX = 800,
+            dimY = 800,
+            internalFilename = "full.jp2",
+            qualityLevel = 100,
+            qualityName = Some("full")
+        ),
+            StillImageFileValueV1(// thumbnail representation
+                internalMimeType = "image/jpeg",
+                originalFilename = originalFilename,
+                originalMimeType = Some(originalMimeType),
+                dimX = 80,
+                dimY = 80,
+                internalFilename = "thumb.jpg",
+                qualityLevel = 10,
+                qualityName = Some("thumbnail"),
+                isPreview = true
+            ))
+
+        deleteTmpFile(conversionRequest)
+
+        Future(SipiResponderConversionResponseV1(fileValuesV1, file_type = SipiConstants.FileType.IMAGE))
+    }
+
+    val sipiMock = actor("mocksipi")(new Act {
+        become {
+            case sipiResponderConversionFileRequest: SipiResponderConversionFileRequestV1 => future2Message(sender(), imageConversionResponse(sipiResponderConversionFileRequest), log)
+            case sipiResponderConversionPathRequest: SipiResponderConversionPathRequestV1 => future2Message(sender(), imageConversionResponse(sipiResponderConversionPathRequest), log)
+        }
+    })
+
+    val responderManager = system.actorOf(Props(new TestResponderManagerV1(Map(SIPI_ROUTER_ACTOR_NAME -> sipiMock))), name = RESPONDER_MANAGER_ACTOR_NAME)
 
     val storeManager = system.actorOf(Props(new StoreManager with LiveActorMaker), name = STORE_MANAGER_ACTOR_NAME)
