+++ webapi/src/test/scala/org/knora/webapi/responders/v1/ResourcesResponderV1Spec.scala
@@ -22,6 +22,7 @@ package org.knora.webapi.responders.v1
 
 import java.util.UUID
 
+import akka.actor.ActorDSL._
 import akka.actor.Props
 import akka.testkit.{ImplicitSender, TestActorRef}
 import org.knora.webapi._
@@ -32,8 +33,11 @@ import org.knora.webapi.messages.v1respondermessages.usermessages.{UserDataV1, U
 import org.knora.webapi.messages.v1respondermessages.valuemessages._
 import org.knora.webapi.responders._
 import org.knora.webapi.store._
+import org.knora.webapi.util.ActorUtil._
 import org.knora.webapi.util._
 
+import scala.concurrent.ExecutionContext.Implicits.global
+import scala.concurrent.Future
 import scala.concurrent.duration._
 
 /**
@@ -66,10 +70,67 @@ object ResourcesResponderV1Spec {
   */
 class ResourcesResponderV1Spec extends CoreSpec() with ImplicitSender {
 
+
+    /**
+      * Imitates the Sipi server by returning a [[SipiResponderConversionResponseV1]] representing an image conversion request.
+      *
+      * @param conversionRequest the conversion request to be handled.
+      * @return a [[SipiResponderConversionResponseV1]] imitating the answer from Sipi.
+      */
+    private def imageConversionResponse(conversionRequest: SipiResponderConversionRequestV1): Future[SipiResponderConversionResponseV1] = {
+
+        // delete tmp file (depending on the kind of request given: only necessary if Knora stored the file - non GUI-case)
+        def deleteTmpFile(conversionRequest: SipiResponderConversionRequestV1): Unit = {
+            conversionRequest match {
+                case (conversionPathRequest: SipiResponderConversionPathRequestV1) =>
+                    // a tmp file has been created by the resources route (non GUI-case), delete it
+                    InputValidation.deleteFileFromTmpLocation(conversionPathRequest.source)
+                case _ => ()
+            }
+        }
+
+        val originalFilename = conversionRequest.originalFilename
+        val originalMimeType: String = conversionRequest.originalMimeType
+
+        val fileValuesV1 = Vector(StillImageFileValueV1(// full representation
+            internalMimeType = "image/jp2",
+            originalFilename = originalFilename,
+            originalMimeType = Some(originalMimeType),
+            dimX = 800,
+            dimY = 800,
+            internalFilename = "full.jp2",
+            qualityLevel = 100,
+            qualityName = Some("full")
+        ),
+            StillImageFileValueV1(// thumbnail representation
+                internalMimeType = "image/jpeg",
+                originalFilename = originalFilename,
+                originalMimeType = Some(originalMimeType),
+                dimX = 80,
+                dimY = 80,
+                internalFilename = "thumb.jpg",
+                qualityLevel = 10,
+                qualityName = Some("thumbnail"),
+                isPreview = true
+            ))
+
+        deleteTmpFile(conversionRequest)
+
+        Future(SipiResponderConversionResponseV1(fileValuesV1, file_type = SipiConstants.FileType.IMAGE))
+    }
+
+    val sipiMock = actor("mocksipi")(new Act {
+        become {
+            case sipiResponderConversionFileRequest: SipiResponderConversionFileRequestV1 => future2Message(sender(), imageConversionResponse(sipiResponderConversionFileRequest), logger)
+            case sipiResponderConversionPathRequest: SipiResponderConversionPathRequestV1 => future2Message(sender(), imageConversionResponse(sipiResponderConversionPathRequest), logger)
+        }
+    })
+
     // Construct the actors needed for this test.
     private val actorUnderTest = TestActorRef[ResourcesResponderV1]
-    //private val responderManager = system.actorOf(Props(new ResponderManagerV1 with LiveActorMaker), name = RESPONDER_MANAGER_ACTOR_NAME)
-    val responderManager = system.actorOf(Props(new TestResponderManagerV1), name = RESPONDER_MANAGER_ACTOR_NAME)
+
+    val responderManager = system.actorOf(Props(new TestResponderManagerV1(Map(SIPI_ROUTER_ACTOR_NAME -> sipiMock))), name = RESPONDER_MANAGER_ACTOR_NAME)
+
     private val storeManager = system.actorOf(Props(new StoreManager with LiveActorMaker), name = STORE_MANAGER_ACTOR_NAME)
 
     val rdfDataObjects = List(
@@ -520,7 +581,7 @@ class ResourcesResponderV1Spec extends CoreSpec() with ImplicitSender {
                 case response: ResourceFullResponseV1 => () // If we got a ResourceFullResponseV1, the operation succeeded.
             }
         }
-        
+
         "create an incunabula:page with a resource pointer" in {
             val recto = TextValueV1("recto")
             val origname = TextValueV1("Blatt")
